Move dataloader fix to hijacks
