+++ infer/modules/ipex/__init__.py
@@ -4,19 +4,8 @@ import contextlib
 import torch
 import intel_extension_for_pytorch as ipex # pylint: disable=import-error, unused-import
 from .hijacks import ipex_hijacks
-from .dataloader import dataloader_init
 
 # pylint: disable=protected-access, missing-function-docstring, line-too-long
-
-#ControlNet depth_leres++
-class DummyDataParallel(torch.nn.Module): # pylint: disable=missing-class-docstring, unused-argument, too-few-public-methods
-    def __new__(cls, module, device_ids=None, output_device=None, dim=0): # pylint: disable=unused-argument
-        if isinstance(device_ids, list) and len(device_ids) > 1:
-            print("IPEX backend doesn't support DataParallel on multiple XPU devices")
-        return module.to("xpu")
-
-def return_null_context(*args, **kwargs): # pylint: disable=unused-argument
-    return contextlib.nullcontext()
 
 def ipex_init(): # pylint: disable=too-many-statements
     #Replace cuda with xpu:
@@ -160,10 +149,7 @@ def ipex_init(): # pylint: disable=too-many-statements
     torch.cuda.get_device_capability = lambda: [11,7]
     torch.cuda.get_device_properties.major = 11
     torch.cuda.get_device_properties.minor = 7
-    torch.backends.cuda.sdp_kernel = return_null_context
-    torch.nn.DataParallel = DummyDataParallel
     torch.cuda.ipc_collect = lambda: None
     torch.cuda.utilization = lambda: 0
 
     ipex_hijacks()
-    dataloader_init()
